hdr: add BaseStoryPart.next_id
